# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: BearMount
    url: https://github.com/WhispersOfJ/bearmount/issues
  description: REST API for BearMount — Usenet WebDAV server with NZB queue,
    health monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: https://github.com/WhispersOfJ/bearmount
  title: BearMount API
  version: "1.0"
paths:
...
  "/config/providers/{id}/tune-pipeline":
    post:
      description: Sweeps inflight depths against the provider and returns the
        recommended inflight_requests value.
      parameters:
        - description: Provider ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.PipelineTuneResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Auto-tune provider pipeline depth
      tags:
        - Providers
...
  /import/nzbdav:
    delete:
      description: Cancels the currently running NZBDav import operation.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Cancel NZBDav import
      tags:
        - Import
    post:
      description: Starts an import from a WebDAV/NZBDav source, fetching NZBs from
        the remote.
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: Import configuration (uses server config if omitted)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Import from NZBDav source
      tags:
        - Import
  /import/nzbdav/migrate-symlinks:
    post:
      description: Walks a library directory and rewrites symlinks that target the
        nzbdav mount to point at the bearmount path instead.
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: library_path, source_mount_path, dry_run
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Migrate NZBDav library symlinks
      tags:
        - Import
  /import/nzbdav/migrations:
    delete:
      description: Deletes every import_migrations row for source='nzbdav' regardless
        of status. Use to force a full re-import after the imported files have
        been deleted from BearMount. This will cause the scanner to re-process
        every blob on the next import.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Clear ALL NZBDav migration rows
      tags:
        - Import
  /import/nzbdav/pending-migrations:
    delete:
      description: Deletes all import_migrations rows with status='pending' for
        source='nzbdav'. Keeps imported/symlinks_migrated rows untouched. Use to
        remove orphaned rows from a cancelled/failed import before re-importing.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Clear pending NZBDav migration rows
      tags:
        - Import
  /import/nzbdav/reset:
    post:
      description: Resets the NZBDav import state so a new import can be started.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Reset NZBDav import status
      tags:
        - Import
  /import/nzbdav/status:
    get:
      description: Returns the current status of the NZBDav import operation.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get NZBDav import status
      tags:
        - Import
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.PipelineDepthSample:
      properties:
        depth:
          type: integer
        mbps:
          type: number
      type: object
    api.PipelineTuneResponse:
      properties:
        baseline_mbps:
          type: number
        best_mbps:
          type: number
        enabled:
          type: boolean
        improvement_pct:
          type: number
        recommended_inflight:
          type: integer
        test_connections:
          type: integer
        tested:
          items:
            $ref: "#/components/schemas/api.PipelineDepthSample"
          type: array
        warning:
          type: string
      type: object
...